LPS-118481 Refactor NewTopic component to extract the modal into a new component
